backend/requirements.txt
pandas
python.dotenv
numpy
langchain
langchain-ollama
